remove compilation warning

+++ src/cpp/rtps/builtin/data/ReaderProxyData.cpp
@@ -59,8 +59,8 @@ ReaderProxyData::ReaderProxyData (
         const VariableLengthDataLimits& data_limits)
     : ReaderProxyData(max_unicast_locators, max_multicast_locators)
 {
-    m_qos.m_userData.max_size(data_limits.max_user_data);
-    m_qos.m_partition.max_size(data_limits.max_partitions);
+    m_qos.m_userData.max_size((uint32_t)data_limits.max_user_data);
+    m_qos.m_partition.max_size((uint32_t)data_limits.max_partitions);
 }
 
 ReaderProxyData::~ReaderProxyData()
@@ -836,8 +836,8 @@ void ReaderProxyData::clear()
     m_isAlive = true;
     m_topicKind = NO_KEY;
     //clear user data but keep max size on qos
-    size_t max_user_data = m_qos.m_userData.max_size();
-    size_t max_partitions = m_qos.m_partition.max_size();
+    uint32_t max_user_data = m_qos.m_userData.max_size();
+    uint32_t max_partitions = m_qos.m_partition.max_size();
     m_qos = ReaderQos();
     m_qos.m_userData.max_size(max_user_data);
     m_qos.m_partition.max_size(max_partitions);
